% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{default_parameters}
\alias{default_parameters}
\title{Useful default parameters}
\usage{
default_parameters(population)
}
\arguments{
\item{population}{An object of class \verb{<population>} from the \pkg{epidemics}
package.}
}
\description{
Useful default parameters
}
