% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noromod-package.R
\docType{package}
\name{noromod-package}
\alias{noromod}
\alias{noromod-package}
\title{noromod: Model a Norovirus Epidemic with Age Structure and Seasonal Forcing}
\description{
Models a norovirus epidemic with age structure and seasonal forcing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pratikunterwegs/noromod}
  \item Report bugs at \url{https://github.com/pratikunterwegs/noromod/issues}
}

}
\author{
\strong{Maintainer}: Pratik Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID})

}
\keyword{internal}
