% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norovirus_model_cpp}
\alias{norovirus_model_cpp}
\title{Compartmental model function for a norovirus epidemic}
\usage{
norovirus_model_cpp(t, state, parameters)
}
\arguments{
\item{t}{Double value for the time.}

\item{state}{Vector of the initial state. The form of the vector should be
\eqn{X_i, for X \in S, E, I_s, I_a, R, I_new, I_re}, where values for each
of the compartments for each age group \eqn{i} are given consecutively.}

\item{parameters}{List for the parameters used in the simulation. See
\code{\link[=default_parameters]{default_parameters()}} for a list of parameters used in development.}
}
\value{
A list with a single numeric vector of the same size as \code{state},
suitable as output for \code{\link[deSolve:lsoda]{deSolve::lsoda()}}.
}
\description{
A compartmental norovirus model with four age groups. This
function is intended to be passed to the \pkg{deSolve} function \code{lsoda}.
}
