% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{output_to_df}
\alias{output_to_df}
\title{Handle data from the Norovirus Boost model}
\usage{
output_to_df(output)
}
\arguments{
\item{output}{An output from the function \code{\link[=noromod_cpp_boost]{noromod_cpp_boost()}}, which is a
list with the elements \code{x} for the model state, and \code{time} for the simulation
time.}
}
\value{
A data.frame with columns named \verb{<compartment>_<age_group>}, and time
for the state at each time point.
}
\description{
Handle data from the Norovirus Boost model
}
