% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noromod_r.R
\name{seasonal_forcing}
\alias{seasonal_forcing}
\title{Seasonal forcing coefficient for the norovirus model}
\usage{
seasonal_forcing(t, w1, w2)
}
\arguments{
\item{t}{Double value for the time, taken as the ordinal day}

\item{w1}{Double value for the first weighting factor}

\item{w2}{Double value for the second weighting factor}
}
\value{
Double value for the coefficient of seasonal forcing
}
\description{
Seasonal forcing coefficient for the norovirus model
}
\keyword{internal}
