% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{default_parameters}
\alias{default_parameters}
\title{Useful default parameters}
\usage{
default_parameters(population)
}
\arguments{
\item{population}{An optional argument taking an object of class
\verb{<population>} from the \pkg{epidemics}. Used to obtain a contact matrix.
package.}
}
\description{
Useful default parameters
}
